requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
soupsieve==2.0.1
Werkzeug==1.0.1
gunicorn
pymongo